urban_proj/urban_assignment/requirements.txt
amqp==2.5.0
APScheduler==3.6.0
billiard==3.6.0.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-apscheduler==0.2.13
django-chartit==0.2.9
django-environ==0.4.5
idna==2.8
Jinja2==2.10.1
joblib==0.13.2
kombu==4.6.3
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
patsy==0.5.1
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
statsmodels==0.10.1
tzlocal==2.0.0
urllib3==1.25.3
vine==1.3.0
